+++ android/src/main/kotlin/com/github/florent37/assets_audio_player/AssetsAudioPlayerPlugin.kt
@@ -2,18 +2,19 @@ package com.github.florent37.assets_audio_player
 
 import StopWhenCall
 import StopWhenCallAudioFocus
+import android.app.Activity
 import android.content.Context
+import android.content.Intent
 import androidx.annotation.NonNull
-import com.github.florent37.assets_audio_player.notification.MediaButtonsReceiver
-import com.github.florent37.assets_audio_player.notification.NotificationManager
-import com.github.florent37.assets_audio_player.notification.fetchAudioMetas
-import com.github.florent37.assets_audio_player.notification.fetchNotificationSettings
+import com.github.florent37.assets_audio_player.notification.*
 import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
-
+import io.flutter.plugin.common.PluginRegistry
 
 internal val METHOD_POSITION = "player.position"
 internal val METHOD_VOLUME = "player.volume"
@@ -28,7 +29,10 @@ internal val METHOD_PREV = "player.prev"
 internal val METHOD_PLAY_OR_PAUSE = "player.playOrPause"
 internal val METHOD_NOTIFICATION_STOP = "player.stop"
 
-class AssetsAudioPlayerPlugin : FlutterPlugin {
+class AssetsAudioPlayerPlugin : FlutterPlugin,PluginRegistry.NewIntentListener, ActivityAware {
+
+    var myActivity: Activity? = null
+    var notificationChannel : MethodChannel ?= null
 
     companion object {
         var instance: AssetsAudioPlayerPlugin? = null
@@ -38,6 +42,7 @@ class AssetsAudioPlayerPlugin : FlutterPlugin {
 
     override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
         instance = this
+        notificationChannel = MethodChannel(flutterPluginBinding.binaryMessenger, "assets_audio_player_notification")
         assetsAudioPlayer = AssetsAudioPlayer(
                 flutterAssets = flutterPluginBinding.flutterAssets,
                 context = flutterPluginBinding.applicationContext,
@@ -49,6 +54,45 @@ class AssetsAudioPlayerPlugin : FlutterPlugin {
     override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
         assetsAudioPlayer?.unregister()
         instance = null
+    }
+
+    private fun sendNotificationPayloadMessage(intent: Intent): Boolean? {
+        if (NotificationAction.ACTION_SELECT == intent.action) {
+            val trackId = intent.getStringExtra(NotificationService.TRACK_ID)
+            notificationChannel?.invokeMethod("selectNotification", trackId)
+            return true
+        }
+        return false
+    }
+
+    override fun onNewIntent(intent: Intent?): Boolean {
+        if (!intent!!.getBooleanExtra("isVisited", false)) {
+            val res = sendNotificationPayloadMessage(intent!!)!!
+            if (res && myActivity != null) {
+                myActivity!!.intent = intent
+                intent.putExtra("isVisited", true)
+            }
+            return res
+        }
+        return false
+    }
+
+    override fun onDetachedFromActivity() {
+        myActivity = null;
+    }
+
+    override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+        binding.addOnNewIntentListener(this);
+        myActivity = binding.getActivity();
+    }
+
+    override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+        binding.addOnNewIntentListener(this);
+        myActivity = binding.getActivity();
+    }
+
+    override fun onDetachedFromActivityForConfigChanges() {
+        myActivity = null;
     }
 }
 
@@ -91,7 +135,7 @@ class AssetsAudioPlayer(
 
     fun unregister() {
         stopWhenCall?.stop()
-        notificationManager.hideNotification(definitively = true)
+        notificationManager.hideNotificationService(definitively = true)
         stopWhenCall?.unregister(stopWhenCallListener)
         players.values.forEach {
             it.stop()
@@ -246,6 +290,23 @@ class AssetsAudioPlayer(
                         return
                     }
                     getOrCreatePlayer(id).setPlaySpeed(speed)
+                    result.success(null)
+                } ?: run {
+                    result.error("WRONG_FORMAT", "The specified argument must be an Map<*, Any>.", null)
+                    return
+                }
+            }
+            "showNotification" -> {
+                (call.arguments as? Map<*, *>)?.let { args ->
+                    val id = args["id"] as? String ?: run {
+                        result.error("WRONG_FORMAT", "The specified argument (id) must be an String.", null)
+                        return
+                    }
+                    val show = args["show"] as? Boolean ?: run {
+                        result.error("WRONG_FORMAT", "The specified argument (show) must be an Boolean.", null)
+                        return
+                    }
+                    getOrCreatePlayer(id).showNotification(show)
                     result.success(null)
                 } ?: run {
                     result.error("WRONG_FORMAT", "The specified argument must be an Map<*, Any>.", null)