+++ lib/src/assets_audio_player.dart
@@ -70,6 +70,45 @@ class AssetsAudioPlayer {
   static final NotificationSettings defaultNotificationSettings =
       const NotificationSettings();
 
+  //region notification click
+  static MethodChannel _notificationOpenChannel =
+      const MethodChannel('assets_audio_player_notification');
+  static final BehaviorSubject<ClickedNotificationWrapper>
+      __onNotificationClicked = BehaviorSubject<ClickedNotificationWrapper>();
+  static final Stream<ClickedNotificationWrapper> _onNotificationClicked =
+      __onNotificationClicked.stream;
+
+  static void setupNotificationsOpenAction(NotificationOpenAction action) {
+    WidgetsFlutterBinding.ensureInitialized();
+    _notificationOpenChannel =
+        const MethodChannel('assets_audio_player_notification');
+    _notificationOpenChannel.setMethodCallHandler((call) async {
+      switch (call.method) {
+        case "selectNotification":
+          {
+            final String audioId = call.arguments;
+            __onNotificationClicked.value =
+                ClickedNotificationWrapper(ClickedNotification(
+              audioId: audioId,
+            ));
+            break;
+          }
+      }
+    });
+    addNotificationOpenAction(action);
+  }
+
+  static StreamSubscription addNotificationOpenAction(
+      NotificationOpenAction action) {
+    return _onNotificationClicked.listen((ClickedNotificationWrapper clicked) {
+      if (action != null && !clicked.handled) {
+        bool handled = action(clicked.clickedNotification);
+        clicked.handled = handled;
+      }
+    });
+  }
+  //endregion
+
   static final uuid = Uuid();
 
   /// The channel between the native and Dart
@@ -286,6 +325,15 @@ class AssetsAudioPlayer {
 
   bool get respectSilentMode => _respectSilentMode;
 
+  bool _showNotification = false;
+  bool get showNotification => _showNotification;
+  set showNotification(bool newValue) {
+    _showNotification = newValue;
+
+    /* await */ _sendChannel.invokeMethod(
+        'showNotification', {"id": this.id, "show": _showNotification});
+  }
+
   /// assign the looping state : true -> looping, false -> not looping
   set loop(value) {
     setLoop(value);
@@ -348,7 +396,7 @@ class AssetsAudioPlayer {
   _init() {
     _recieveChannel = MethodChannel('assets_audio_player/$id');
     _recieveChannel.setMethodCallHandler((MethodCall call) async {
-      print("received call ${call.method} with arguments ${call.arguments}");
+      //print("received call ${call.method} with arguments ${call.arguments}");
       switch (call.method) {
         case 'log':
           print("log: " + call.arguments);
@@ -501,6 +549,7 @@ class AssetsAudioPlayer {
       this.isShuffling,
       this.current,
       this.currentPosition,
+      this.isBuffering
     ])
         .map((values) => RealtimePlayingInfos(
               volume: values[0],
@@ -509,6 +558,7 @@ class AssetsAudioPlayer {
               isShuffling: values[3],
               current: values[4],
               currentPosition: values[5],
+              isBuffering: values[6],
               playerId: this.id,
             ))
         .listen((readingInfos) {
@@ -685,6 +735,7 @@ class AssetsAudioPlayer {
     final currentAudio = _lastOpenedAssetsAudio;
     if (audioInput != null) {
       _respectSilentMode = respectSilentMode;
+      _showNotification = showNotification;
 
       final audio = await _handlePlatformAsset(audioInput);